requirements.txt
pyserial
pyodbc